requirements.txt
apscheduler==3.7.0
asgiref==3.4.1
channels==3.0.4
channels_redis==3.3.0
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jsonapi==4.3.0
django-apscheduler==0.6.0
django-environ==0.4.5
django_cors_headers==3.8.0
django-dramatiq==0.10.0
django-filter==2.4.0
django-netfields==1.2.2
django-redis==5.0.0
django-qr-code==2.2.0
drf-nested-routers>=0.92.5
django-phonenumber-field[phonenumberslite]==5.2.0
dramatiq==1.11.0
drf-yasg-json-api==0.8.1
drf-yasg==1.20.0
gunicorn==20.1.0
netaddr==0.8.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
rollbar==0.16.1
sqlparse==0.4.1
spyne==2.13.16
soapfish2==0.7.1
transitions==0.8.8
zeep==4.1.0
httpx==0.19.0
drf-renderer-xlsx==0.4.1
